requirements.txt
absl-py==0.12.0
altair==4.1.0
argon2-cffi==20.1.0
astor==0.8.1
astroid==2.5.1
astunparse==1.6.3
async-generator==1.10
attrs==21.2.0
autocrop==1.1.1
backcall==0.2.0
base58==2.1.0
bleach==3.3.0
blinker==1.4
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
cycler==0.10.0
decorator==5.0.9
defusedxml==0.7.1
entrypoints==0.3
flatbuffers==1.12
gast==0.4.0
gitdb==4.0.7
GitPython==3.1.18
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.34.1
h5py==3.1.0
idna==2.10
ipykernel==5.5.5
ipython==7.24.1
ipython-genutils==0.2.0
ipywidgets==7.6.3
isort==5.8.0
jedi==0.18.0
Jinja2==3.0.1
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
lazy-object-proxy==1.6.0
Markdown==3.3.4
MarkupSafe==2.0.1
mccabe==0.6.1
mistune==0.8.4
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.0
numpy==1.20.3
olefile==0.46 
opencv-python==4.5.2.54
opt-einsum==3.3.0
packaging==20.9
pandas==1.2.4
pandocfilters==1.4.3
parso==0.8.2
pickleshare==0.7.5
Pillow==8.1.2
prometheus-client==0.11.0
prompt-toolkit==3.0.19
protobuf==3.17.3
pycparser==2.20
pydeck==0.6.2
Pygments==2.9.0
pylint==2.7.2
pyparsing==2.4.7
PyQt5==5.12.3
PyQt5-sip==4.19.18
PyQtChart==5.12
PyQtWebEngine==5.12.1
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2021.1
pywin32==301
pywinpty==1.1.2
pyzmq==22.1.0
requests==2.25.1
Send2Trash==1.5.0
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
smmap==4.0.0
streamlit==0.83.0
tensorboard==2.5.0
tensorflow==2.5.0
tensorflow-addons==0.13.0
tensorflow-estimator==2.5.0
termcolor==1.1.0
terminado==0.10.1
testpath==0.5.0
toml==0.10.2
toolz==0.11.1
tornado @ file:///D:/bld/tornado_1610094910414/work
traitlets==5.0.5
typeguard==2.12.1
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.26.5
validators==0.18.2
watchdog==2.1.2
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
wincertstore==0.2
wrapt==1.12.1
